Generalized WebRTX Control and Step calls from omnidreams and lingobot
