+++ integrations/omnidreams/omnidreams/webrtc/server.py
@@ -4,8 +4,6 @@
 from __future__ import annotations
 
 import argparse
-import gc
-import logging
 import os
 from pathlib import Path
 
@@ -21,22 +19,17 @@ from omnidreams.webrtc.session import (
 )
 
 from flashdreams.core.distributed import (
-    configure_loguru_for_distributed,
-)
-from flashdreams.core.distributed import (
     init as distributed_init,
 )
 from flashdreams.serving.network import get_external_ip
+from flashdreams.serving.webrtc.bootstrap import (
+    configure_logging,
+    run_webrtc_server,
+)
 from flashdreams.serving.webrtc.server import WebRTCSessionManager, create_webrtc_app
 
 WEB_DIR = Path(__file__).resolve().parent / "web"
 REPO_ASSETS_DIR = Path(__file__).resolve().parents[4] / "assets"
-
-
-def configure_logging(*, world_rank: int | None = None) -> None:
-    configure_loguru_for_distributed(world_rank=world_rank)
-    for logger_name in ("aioice", "aioice.ice", "aiortc"):
-        logging.getLogger(logger_name).setLevel(logging.WARNING)
 
 
 def parse_args() -> argparse.Namespace:
@@ -224,6 +217,7 @@ def main() -> None:
     runtime_device, world_rank, _ = initialize_distributed(default_device=args.device)
     runtime_config = build_runtime_config(args, device_override=str(runtime_device))
     session_manager = OmnidreamsWebRTCSessionManager(runtime_config=runtime_config)
+    app = None
     if world_rank == 0:
         external_ip = get_external_ip()
         app = create_app(
@@ -231,24 +225,13 @@ def main() -> None:
             request_session_url=f"http://{external_ip}:{args.port}/request_session",
         )
         logger.info("Starting on external IP: {}", external_ip)
-        try:
-            web.run_app(app, host=args.host, port=args.port)
-        finally:
-            session_manager.send_exit_signal()
-    else:
-        try:
-            session_manager.wait_for_termination()
-        except KeyboardInterrupt:
-            logger.warning("Worker rank interrupted, shutting down.")
-
-    gc.collect()
-    if torch.cuda.is_available():
-        torch.cuda.empty_cache()
-        torch.cuda.synchronize()
-
-    if dist.is_initialized():
-        dist.barrier()
-        dist.destroy_process_group()
+    run_webrtc_server(
+        world_rank=world_rank,
+        session_manager=session_manager,
+        app=app,
+        host=args.host,
+        port=args.port,
+    )
 
 
 if __name__ == "__main__":
